# openapi.yml
swagger: "2.0"
info:
  description: "This is an API to scan docker images for security vulnerabilities. The source code is available in [GitHub](https://github.com/abhisek/container-image-scanner-api)"
  version: "1.0.0"
  title: "Container Image Scanner API"
  license:
    name: "MIT"
    url: "https://github.com/abhisek/container-image-scanner-api/blob/master/LICENSE"
host: "public-api-services.kubectl.in"
basePath: "/container-image-scanner-api/v1"
tags:
- name: "scan"
  description: "A Scanner Job"
schemes:
- "https"
paths:
  /scans:
    post:
      tags:
      - "scan"
      summary: "Submit new scan"
      description: ""
      operationId: "submitScan"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Scan object that needs to be submitted for scan"
        required: true
        schema:
          $ref: "#/definitions/ScanRequest"
      responses:
        405:
          description: "Invalid input"
  /scans/{scanId}:
    get:
      description: "Returns a single scan"
      produces:
      - "application/json"
      parameters:
      - name: "scanId"
        in: "path"
        description: "ID of scan to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ScanResponse"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Scan not found"
  /scans/{scanId}/status:
    post:
      tags:
      - "scan"
      summary: "Check scan status"
      produces:
      - "application/json"
      parameters:
      - name: "scanId"
        in: "path"
        description: "ID of scan"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ScanStatus"
definitions:
  ScanResponse:
    type: "object"
    properties:
      id:
        type: "string"
      vulnerabilities:
        type: object
      audit:
        type: object
  ScanStatus:
    type: "object"
    properties:
      status:
        type: "string"
        enum:
          - NEW
          - IN-PROGRESS
          - COMPLETED
          - ERROR
  ScanRequest:
    type: "object"
    properties:
      image:
        type: "string"
        description: "The image to be scanned"
      username:
        type: "string"
        description: "Registry authentication (username)"
      password:
        type: "string"
        description: "Registry authentication (password)"